bugfix: themeContext keep reverting to darkTheme after page reload
